+++ FlutterWave.Core/Models/Services/Foundations/FlutterWave/Transfers/TransferRatesResponse.cs
@@ -16,13 +16,13 @@
         public class Destination
         {
             public string Currency { get; set; }
-            public int Amount { get; set; }
+            public decimal Amount { get; set; }
         }
 
         public class Source
         {
             public string Currency { get; set; }
-            public int Amount { get; set; }
+            public decimal Amount { get; set; }
         }
